Move `Scalar` into `crypto` module, the only change is changed visibility for its internal field
