+++ crates/subspace-erasure-coding/src/tests.rs
@@ -1,7 +1,7 @@
 use crate::ErasureCoding;
 use std::iter;
 use std::num::NonZeroUsize;
-use subspace_core_primitives::Scalar;
+use subspace_core_primitives::crypto::Scalar;
 
 // TODO: This could have been done in-place, once implemented can be exposed as a utility
 fn concatenated_to_interleaved<T>(input: Vec<T>) -> Vec<T>
