+++ crates/subspace-farmer/src/identity.rs
@@ -5,7 +5,8 @@ use schnorrkel::{ExpansionMode, Keypair, PublicKey, SecretKey, Signature};
 use std::fs;
 use std::ops::Deref;
 use std::path::Path;
-use subspace_core_primitives::{ChunkSignature, Scalar};
+use subspace_core_primitives::crypto::Scalar;
+use subspace_core_primitives::ChunkSignature;
 use subspace_solving::{create_chunk_signature, REWARD_SIGNING_CONTEXT};
 use substrate_bip39::mini_secret_from_entropy;
 use tracing::debug;
